// src/types/personalization/typesColorScales.ts
import { CodesForColorScales } from './enumsPersonalization';

export type ScalesForColors = {
  50: string;
  100: string;
  200: string;
  300: string;
  400: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
  950: string;
};

export type ColorScales = {
  id: CodesForColorScales;
  name: string;
  scales: ScalesForColors;
};